+++ src/include/Platform/StringRequireAutoImporter.hpp
@@ -24,6 +24,7 @@ using AliasMap = DenseHashMap<std::string, Luau::Config::AliasInfo>;
 
 std::string requireNameFromModuleName(const Luau::ModuleName& name);
 std::optional<std::string> computeBestAliasedPath(const Uri& to, const AliasMap& availableAliases);
-std::pair<std::string, SortText::SortTextT> computeRequirePath(const Uri& from, const Uri& to, const AliasMap& availableAliases, ImportRequireStyle importRequireStyle);
+std::pair<std::string, SortText::SortTextT> computeRequirePath(
+    const Uri& from, Uri to, const AliasMap& availableAliases, ImportRequireStyle importRequireStyle);
 void suggestStringRequires(const StringRequireAutoImporterContext& ctx, std::vector<lsp::CompletionItem>& items);
 }