+++ src/platform/StringRequireAutoImporter.cpp
@@ -7,6 +7,10 @@ namespace Luau::LanguageServer::AutoImports
 std::string requireNameFromModuleName(const Luau::ModuleName& name)
 {
     auto fileName = name;
+
+    if (isInitLuauFile(name))
+        fileName = getParentPath(name).value_or(name);
+
     if (const auto slashPos = fileName.find_last_of('/'); slashPos != std::string::npos)
         fileName =  fileName.substr(slashPos + 1);
     fileName = removeSuffix(fileName, ".luau");
@@ -45,9 +49,13 @@ std::optional<std::string> computeBestAliasedPath(const Uri& to, const AliasMap&
 }
 
 /// Resolves to the most appropriate style of string require based off configuration / heuristics
-std::pair<std::string, SortText::SortTextT> computeRequirePath(const Uri& from, const Uri& to, const AliasMap& availableAliases, ImportRequireStyle importRequireStyle)
+std::pair<std::string, SortText::SortTextT> computeRequirePath(
+    const Uri& from, Uri to, const AliasMap& availableAliases, ImportRequireStyle importRequireStyle)
 {
     auto fromParent = from.parent();
+
+    if (isInitLuauFile(to.fsPath()))
+        to = to.parent().value_or(to);
 
     if (importRequireStyle != ImportRequireStyle::AlwaysRelative)
     {